# docs/static/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Singularity API
  description: "Welcome to the *Singularity*! \n\nThis documentation covers the ready-to-use\
    \ endpoints \nfor authentication, user management, and content management that\
    \ come pre-configured with \n*Singularity*.\n\n**Features at a Glance**:\n\n-\
    \ **Authentication & User Management:** Includes JWT auth with refresh tokens,\
    \ 2FA, \n  email verification, and role-based access control.\n  \n- **Data &\
    \ Content Management:** Endpoints for interacting with a content management \n\
    \  system that supports multi-language content, configurable tagging, and built-in\
    \ \n  access control. A prebuilt `Article` class is also provided.\n  \n- **Performance\
    \ & Security:** The API is built with a focus on production-readiness, \n  including\
    \ rate limiting and secret manager integration with automated key rotation.\n\n\
    This API is a foundation for building robust backend services, microservices,\
    \ or \nfull-stack applications.\n\n**Note:** Some endpoints require explicit configuration\
    \ to be enabled."
  contact:
    name: Stereov
    url: https://stereov.io
    email: contact@stereov.io
  license:
    name: GPLv3 License
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0
servers:
- url: https://singularity.stereov.io
  description: Demo server
- url: http://localhost:8000
  description: Local development server
...
paths:
...
  /api/auth/2fa/email/send:
    post:
      tags:
      - Two-Factor Authentication
      summary: Send Email 2FA Code
      description: "Send an email containing a 2FA code to the user.\n\nLearn more\
        \ about email as 2FA method [here](https://singularity.stereov.io/docs/guides/auth/two-factor#email).\n\
        \nThis endpoint is used to send a 2FA email code for enabling email as a 2FA\
        \ method\nand to resend 2FA codes when authenticating user that already enabled\
        \ email as a 2FA method.\n\n**Note:** Each request will generate a new code\
        \ and invalidate all old codes.\n\nIf email is the preferred 2FA method, an\
        \ email will be sent automatically after successful authentication\nwith the\
        \ user's password.\nYou can learn more about preferred 2FA methods [here](https://singularity.stereov.io/docs/guides/auth/two-factor#changing-the-preferred-method).\n\
        \n**Requirements:**\n- The user can authenticate using password. 2FA will\
        \ not work with OAuth2. \n  The OAuth2 provider will validate the second factor\
        \ if the user enabled it for the provider.\n\n**Tokens:**\n\nThere are two\
        \ options: \n1. If email as a 2FA method is disabled, you can request an email\
        \ with a token to enable it.\n   In this case you need a valid [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token).\n\
        \   If email is already enabled, this request will return `400 - BAD REQUEST`.\n\
        2. If email as 2FA method is already enabled, you can request a 2FA email\
        \ code with a valid\n   [`TwoFactorAuthenticationToken`](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token).\n\
        \   \n**Note:** After each email, a cooldown will be started.\nWhen the cooldown\
        \ is active, no new email can be sent.\nYou can request the remaining cooldown\
        \ throw the endpoint \n[`GET /api/auth/2fa/email/cooldown`](https://singularity.stereov.io/docs/api/get-remaining-email-two-factor-cooldown).\n\
        The cooldown can be configured [here](https://singularity.stereov.io/docs/guides/email/configuration)."
      externalDocs:
        url: https://singularity.stereov.io/docs/guides/auth/two-factor#sending-a-2fa-code-via-email
      operationId: sendEmailTwoFactorCode
      parameters:
      - name: locale
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Success.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: 2FA cannot be sent for users who didn't configure authentication
            using a password.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired `AccessToken` or `TwoFactorAuthenticationToken`.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Cannot send email while cooldown is active.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Email is disabled in your application.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenHeader: []
      - accessTokenCookie: []
      - twoFactorAuthenticationTokenHeader: []
      - twoFactorAuthenticationTokenCookie: []
  /api/auth/2fa/email/enable:
    post:
      tags:
      - Two-Factor Authentication
      summary: Enable Email as 2FA Method
      description: "Enable email as 2FA method.\n\nLearn more about email as 2FA method\
        \ [here](https://singularity.stereov.io/docs/guides/auth/two-factor#email).\n\
        \n**Requirements:**\n- The user can authenticate using password. 2FA will\
        \ not work with OAuth2. \n  The OAuth2 provider will validate the second factor\
        \ if the user enabled it for the provider.\n\n**Note:** If [email is enabled](https://singularity.stereov.io/docs/guides/email/configuration)\
        \ in your application,\nemail is 2FA method will be enabled by default for\
        \ every user that registers with a password.\n\n**Tokens:**\n- A valid [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)\
        \ is required.\n- A valid [`StepUpToken`](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)\n\
        \  is required. This token should match user and session contained in the\
        \ `AccessToken`."
      externalDocs:
        url: https://singularity.stereov.io/docs/guides/auth/two-factor#setup-1
      operationId: enableEmailAsTwoFactorMethod
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableEmailTwoFactorMethodRequest"
        required: true
      responses:
        "200":
          description: Updated user information.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "304":
          description: This method is already enabled for the user.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: 2FA cannot be enabled for users who didn't configure authentication
            using a password.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid `AccessToken` or `StepUpToken`.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenHeader: []
        stepUpTokenHeader: []
      - accessTokenCookie: []
        stepUpTokenCookie: []
  /api/auth/2fa/email:
    delete:
      tags:
      - Two-Factor Authentication
      summary: Disable Email as 2FA Method
      description: |-
        Disable email as 2FA method.

        Learn more about email as 2FA method [here](https://singularity.stereov.io/docs/guides/auth/two-factor#email).

        **Tokens:**
        - A valid [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token) is required.
        - A valid [`StepUpToken`](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)
          is required. This token should match user and session contained in the `AccessToken`.
      externalDocs:
        url: https://singularity.stereov.io/docs/guides/auth/two-factor#disable-1
      operationId: disableEmailAsTwoFactorMethod
      responses:
        "200":
          description: Updated user information.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: This method is already disabled for the user.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid `AccessToken` or `RefreshToken`.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenHeader: []
        stepUpTokenHeader: []
      - accessTokenCookie: []
        stepUpTokenCookie: []
  /api/auth/2fa/email/cooldown:
    get:
      tags:
      - Two-Factor Authentication
      summary: Get Remaining Email 2FA Code Cooldown
      description: |-
        Get the remaining time in seconds until you can send another email containing a 2FA code.

        Learn more about email as 2FA method [here](https://singularity.stereov.io/docs/guides/auth/two-factor#email).

        **Tokens:**
        - Requires a valid [`AccessToken`](https://singularity.stereov.io/docs/guides/auth/tokens#access-token).
      externalDocs:
        url: https://singularity.stereov.io/docs/guides/auth/two-factor#sending-a-2fa-code-via-email
      operationId: getRemainingEmailTwoFactorCooldown
      responses:
        "200":
          description: The remaining cooldown.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MailCooldownResponse"
        "401":
          description: Invalid or expired `AccessToken`.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenHeader: []
      - accessTokenCookie: []
...
components:
  schemas:
...
    ContentAccessDetails:
      type: object
      properties:
        ownerId:
          $ref: "#/components/schemas/ObjectId"
        visibility:
          type: string
          enum:
          - PRIVATE
          - PUBLIC
          - SHARED
        users:
          $ref: "#/components/schemas/ContentAccessPermissions"
        groups:
          $ref: "#/components/schemas/ContentAccessPermissions"
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/ObjectId"
          uniqueItems: true
      required:
      - groups
      - invitations
      - ownerId
      - users
      - visibility
    ContentAccessPermissions:
      type: object
      properties:
        admin:
          type: array
          items:
            type: string
          uniqueItems: true
        editor:
          type: array
          items:
            type: string
          uniqueItems: true
        viewer:
          type: array
          items:
            type: string
          uniqueItems: true
        isEmpty:
          type: boolean
      required:
      - admin
      - editor
      - isEmpty
      - viewer
    FileMetadataResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        key:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        access:
          $ref: "#/components/schemas/ContentAccessDetails"
        contentType:
          type: string
        url:
          type: string
        size:
          type: integer
          format: int64
        trusted:
          type: boolean
        tags:
          type: array
          items:
            type: string
          uniqueItems: true
      required:
      - access
      - contentType
      - createdAt
      - id
      - key
      - size
      - tags
      - trusted
      - updatedAt
      - url
    ObjectId:
      type: object
      properties:
        timestamp:
          type: integer
          format: int32
        date:
          type: string
          format: date-time
    UserResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        name:
          type: string
        email:
          type: string
        identityProviders:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            type: string
            enum:
            - USER
            - ADMIN
            - GUEST
          uniqueItems: true
        emailVerified:
          type: boolean
        lastActive:
          type: string
        twoFactorAuthEnabled:
          type: boolean
        preferredTwoFactorMethod:
          type: string
          enum:
          - totp
          - mail
        twoFactorAuthMethods:
          type: array
          items:
            type: string
            enum:
            - totp
            - mail
        avatar:
          $ref: "#/components/schemas/FileMetadataResponse"
        created:
          type: string
        groups:
          type: array
          items:
            type: string
          uniqueItems: true
      required:
      - created
      - emailVerified
      - groups
      - id
      - identityProviders
      - lastActive
      - name
      - roles
      - twoFactorAuthEnabled
      - twoFactorAuthMethods
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required:
      - error
      - status
      - timestamp
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
      - success
...
    EnableEmailTwoFactorMethodRequest:
      type: object
      properties:
        code:
          type: string
      required:
      - code
...
    MailCooldownResponse:
      type: object
      properties:
        remaining:
          type: integer
          format: int64
      required:
      - remaining
...
  securitySchemes:
    twoFactorAuthenticationTokenCookie:
      type: apiKey
      description: "Token for successful authentication with email and password, indicating\
        \ 2FA is required.You can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token)."
      name: two_factor_authentication_token
      in: cookie
    accessTokenHeader:
      type: http
      description: "Access token for user authentication. You can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)."
      scheme: bearer
      bearerFormat: JWT
    accessTokenCookie:
      type: apiKey
      description: "Access token for user authentication.You can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#access-token)."
      name: access_token
      in: cookie
    stepUpTokenCookie:
      type: apiKey
      description: "Token for step up authentication allowing access of secure resources.You\
        \ can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)."
      name: step_up_token
      in: cookie
    twoFactorAuthenticationTokenHeader:
      type: apiKey
      description: "Token for successful authentication with email and password, indicating\
        \ 2FA is required.You can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#two-factor-authentication-token)."
      name: X-Two-Factor-Authentication-Token
      in: header
    refreshTokenHeader:
      type: http
      description: "Refresh token to request new access tokens.You can learn more\
        \ [here](https://singularity.stereov.io/docs/guides/auth/tokens#refresh-token)."
      scheme: bearer
      bearerFormat: JWT
    stepUpTokenHeader:
      type: apiKey
      description: "Token for step up authentication allowing access of secure resources.You\
        \ can learn more [here](https://singularity.stereov.io/docs/guides/auth/tokens#step-up-token)."
      name: X-Step-Up-Token
      in: header
...